test: clean visualization spy imports
