+++ src/stable_yield_lab/visualization/__init__.py
@@ -1,0 +1,5 @@
+"""Visualization utilities for StableYieldLab."""
+
+from .visualizer import Visualizer
+
+__all__ = ["Visualizer"]
